requirements.txt
PyYAML>=6.0
click>=8.0
python-dotenv>=1.0
pyyaml